refactor(src): :pill: ロジックを別ファイルに退避
    index.jsは処理の結果を返すだけの薄いラッパーにするべきなので
    メインの処理をcommon.jsに全て移行
    ロジックは機能単位ごとに別関数に切り出した

    BREAKING CHANGE: 構成管理するソースファイル名と構成の変更
